requirements.txt
# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
python-multipart==0.0.6

# Google Cloud / Vertex AI
google-cloud-aiplatform>=1.50.0

# RAG dependencies
faiss-cpu>=1.7.4
numpy>=1.24.0

# Datadog observability
datadog==0.48.0

# GitHub API
requests>=2.31.0

# Utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
httpx==0.26.0

# BigQuery
google-cloud-bigquery>=3.13.0
